subscribes and friends

+++ connectapp/friends/views.py
@@ -1,17 +1,16 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.contrib.auth.decorators import login_required
-from django.db.models import Q
 from users.models import User
 from .forms import SearchForm
 from django.core.paginator import Paginator
 from .models import Subscription
-from django.http import HttpResponseForbidden
-from django.contrib.auth import get_user_model
+from django.db.models import Case, When, Value, BooleanField
 
 @login_required
 def friends(request):
     return render(request, "friends/friends.html")
 
+@login_required
 def search_friends(request):
     query = request.GET.get('query', '')
     results = User.objects.filter(
@@ -21,32 +20,21 @@ def search_friends(request):
     ) | User.objects.filter(
         username__icontains=query
     ) if query else User.objects.none()
-    
+    results = results.annotate( is_following=Case( When(followers__user=request.user, then=Value(True)), default=Value(False), output_field=BooleanField(),))
     paginator = Paginator(results, 10)
     page_number = request.GET.get('page')
     page_obj = paginator.get_page(page_number)
 
     form = SearchForm(initial={'query': query})
+    is_following = request.user.followers.filter(user=request.user).exists()
+
     context = {
+        'is_following': is_following,
         'form': form,
         'page_obj': page_obj,
         'query': query,
     }
     return render(request, 'friends/friends.html', context)
-
-@login_required
-def subscribe(request, user_id):
-    user_to_subscribe = get_object_or_404(get_user_model(), id=user_id)
-
-    if user_to_subscribe == request.user:
-        return HttpResponseForbidden("Нельзя подписаться на самого себя.")
-    
-    if not Subscription.objects.filter(user=request.user, subscribed_to=user_to_subscribe).exists():
-        Subscription.objects.create(user=request.user, subscribed_to=user_to_subscribe)
-
-    return redirect('search_friends')
-
-User = get_user_model()
 
 @login_required
 def follow_user(request, username):
@@ -60,20 +48,7 @@ def follow_user(request, username):
 def unfollow_user(request, username):
     """Отписаться от пользователя"""
     user_to_unfollow = get_object_or_404(User, username=username)
-    Subscription.objects.filter(user=request.user, subscribed_to=user_to_unfollow).delete()
+    if user_to_unfollow!= request.user:
+        Subscription.objects.filter(user=request.user, subscribed_to=user_to_unfollow).delete()
     return redirect('user_profile', username=username)
 
-@login_required
-def user_profile(request, username):
-    """Отображение профиля пользователя"""
-    user = get_object_or_404(User, username=username)
-    followers_count = user.followers.count()
-    following_count = user.subscriptions.count()
-    is_following = user.followers.filter(user=request.user).exists()
-
-    return render(request, 'person/user_profile.html', {
-        'user': user,
-        'followers_count': followers_count,
-        'following_count': following_count,
-        'is_following': is_following,
-    })
